+++ src/main/java/cl/duoc/demoJPA006D/repository/DoctorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DoctorRepository extends JpaRepository<Doctor, Integer> {
+public interface DoctorRepository extends JpaRepository<Doctor, Long> {
     List<Doctor> findByName(String name);
     List<Doctor> findByNameAndLastName(String name, String lastName);
 }
