+++ src/main/java/cl/duoc/demoJPA006D/service/PatientService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface PatientService {
     List<PatientDto> findAll();
-    PatientDto findById(int id);
+    PatientDto findById(Long id);
     PatientDto create(PatientDto patient);
     PatientDto update(PatientDto patient);
-    boolean delete(int id);
+    boolean delete(Long id);
 }
